# Optimize `new_data_frame()` and add `new_bare_data_frame()`