--- a/NEWS.md
+++ b/NEWS.md
@@ -1,6 +1,8 @@
 
 # vctrs (development version)
 
+* New `new_bare_data_frame()` is an extremely fast data frame constructor.
+
 * `vec_ptype()` now errors on scalar inputs (#807).
 
 * `vec_ptype_finalise()` is now recursive over all data frame types, ensuring