--- a/R/type-data-frame.R
+++ b/R/type-data-frame.R
@@ -16,7 +16,7 @@
 #' @examples
 #' new_data_frame(list(x = 1:10, y = 10:1))
 new_data_frame <- function(x = list(), n = NULL, ..., class = character()) {
-  .Call(vctrs_new_data_frame, x, n, pairlist(...), class)
+  .Call(vctrs_new_data_frame, x, n, list(...), class)
 }
 
 # Light weight constructor used for tests - avoids having to repeatedly do