--- a/R/type-data-frame.R
+++ b/R/type-data-frame.R
@@ -15,7 +15,7 @@
 #' @keywords internal
 #' @examples
 #' new_data_frame(list(x = 1:10, y = 10:1))
-new_data_frame <- function(x = list(), n = NULL, ..., class = character()) {
+new_data_frame <- function(x = list(), n = NULL, ..., class = character(0)) {
   .Call(vctrs_new_data_frame, x, n, list(...), class)
 }
 new_data_frame <- fn_inline_formals(new_data_frame, c("x", "n", "class"))