--- a/tests/testthat/test-type-data-frame.R
+++ b/tests/testthat/test-type-data-frame.R
@@ -249,3 +249,41 @@ test_that("attributes must be named", {
   expect_error(new_data_frame(list(), n = 0L, 1), "supplied in `...` must be named")
 })
 
+# new_bare_data_frame -----------------------------------------------------
+
+test_that("can construct an empty data frame", {
+  expect_identical(new_bare_data_frame(), data.frame())
+})
+
+test_that("can validly set the number of rows when there are no columns", {
+  expect <- structure(
+    list(),
+    class = "data.frame",
+    row.names = .set_row_names(2L),
+    names = character()
+  )
+
+  expect_identical(new_bare_data_frame(n = 2L), expect)
+})
+
+test_that("size is pulled from first column if not supplied", {
+  x <- new_bare_data_frame(list(x = 1:5, y = 1:6))
+  expect_identical(.row_names_info(x, type = 1), -5L)
+})
+
+test_that("can construct a data frame without column names", {
+  expect_named(new_bare_data_frame(list(1, 2)), NULL)
+})
+
+test_that("the names on an empty data frame are an empty character vector", {
+  expect_identical(names(new_bare_data_frame()), character())
+})
+
+test_that("`x` must be a list", {
+  expect_error(new_bare_data_frame(1), "`x` must be a list")
+})
+
+test_that("if supplied, `n` must be an integer of size 1", {
+  expect_error(new_bare_data_frame(n = c(1L, 2L)), "must be an integer of size 1")
+  expect_error(new_bare_data_frame(n = "x"), "must be an integer of size 1")
+})