It ends up being slightly faster to use `list()`, not `pairlist()`, here. I don't think using `pairlist()` has any advantage

I also inlined the formals for `new_data_frame()`. That helps too.

``` r
bench::mark(
  list(),
  pairlist(),
  check = FALSE,
  iterations = 100000
)
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 list()         20ns     24ns 28975423.        0B       0 
#> 2 pairlist()    478ns    606ns  1507266.    4.58KB     106.
```

``` r
library(vctrs)

# before
bench::mark(
  new_data_frame(),
  iterations = 1000000
)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame()   1.44µs   1.71µs   537193.    9.39KB     66.6

# after `pairlist()` -> `list()` change
bench::mark(
  new_data_frame(),
  iterations = 1000000
)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame()   1.05µs    1.3µs   690766.    4.77KB     76.7

# after additional inlining of formals
bench::mark(
  new_data_frame(),
  iterations = 1000000
)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame()    613ns    744ns  1244202.    4.47KB     114.
```