@DavisVaughan 2020-02-18
This also adds and exports `new_bare_data_frame()`, an extremely fast data frame constructor when we don't care about attributes at all.

``` r
library(vctrs)

bench::mark(
  new_data_frame(),
  new_bare_data_frame(),
  iterations = 1000000
)
#> # A tibble: 2 x 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame()         641ns    864ns   972642.    4.47KB     91.4
#> 2 new_bare_data_frame()    443ns    518ns  1579752.    3.79KB     23.7

bench::mark(
  new_data_frame(mtcars),
  new_bare_data_frame(mtcars),
  iterations = 1000000
)
#> # A tibble: 2 x 6
#>   expression                       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame(mtcars)         847ns   1.26µs   744141.    13.3KB     20.1
#> 2 new_bare_data_frame(mtcars)    628ns    846ns  1086455.        0B     23.9
```